plugins/parse_bilibili/requirements.txt
# B站内容解析插件依赖项
# 基础依赖
requests>=2.28.0
aiohttp>=3.8.0
ujson>=5.4.0
aiofiles>=0.8.0

# B站API相关
bilibili-api-python>=16.0.0
tenacity>=8.0.0

# 图像处理
pillow>=9.0.0
qrcode>=7.3.1

# 模板渲染
jinja2>=3.0.0
bs4>=0.0.1
nonebot-plugin-htmlrender>=0.2.0

# 进度显示
tqdm>=4.64.0

# NoneBot相关依赖
nonebot-plugin-alconna>=0.30.0
nonebot-plugin-session>=0.2.0
nonebot-plugin-uninfo>=0.1.0

# 其他工具
pretty-errors>=1.2.25

# 注意：此插件还需要安装FFmpeg并添加到系统PATH
# Windows: 从 https://ffmpeg.org/download.html 下载
# macOS: brew install ffmpeg
# Linux: apt install ffmpeg 或 yum install ffmpeg
